Loading tasks from Local Storage
Retrieving tasks from the local storage using the 'getItem' method and turning them into a array with 'JSON.parse' method and stored it in a variable called 'tasksFromLocalStorage'. Then wrote an if statement to check if 'tasksFromLocalStorage' is truthy, if so, the tasks are assigned to the 'tasks' array and then the 'renderTasks' function is called. This ensures that any tasks added by the user are persisted even when they reload or close and reopen the browser.